requirements.txt
Click==7.0
dateutils==0.6.7
Flask==1.1.1
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pymongo==3.9.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
tornado==6.0.3
Werkzeug==0.16.0
